reflect: prevent additional StructOf embedded method cases
The current implementation does not wrappers for methods of embedded
non-interface types. We can only skip the wrapper if kindDirectIface
of the generated struct type matches kindDirectIface of the embedded
type. Panic if that is not the case.

It would be better to actually generate wrappers, but that can be done
later.

Updates #15924
Fixes #24782

Change-Id: I01f5c76d9a07f44e1b04861bfe9f9916a04e65ca