Update patch set 5
Patch Set 5: Code-Review+1

> Patch Set 5:
> 
> Peter, I made a small change to the signature type values, and added a sentinel test for PSS certificates. Let me know if it looks good to you.

Thank you, that test looks good.

As for the sentinel values, isn't it misleading to suggest a relation of the values to those in the RFC? That should no longer be the case (hence the use of "iota").

Patch-set: 5
Reviewer: Gerrit User 24681 <24681@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1