Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The current code before this CL has a simple structure with a single loop.  This CL introduces a nested loop.  I can't see any reason to have a nested loop here.  This doesn't seem like the right approach.

The issue has to do with a path that has a symlink that points to another symlink as the setup. If you resolve path1/.., like my example, the current code resolves the first symlink. The Join() comes in to append '..' but the hidden Clean() walks up one segment of the path. When we try to resolve the next round, we are no longer pointing to the symlink but its parent which is where the bug lies.

My other approach that also works is to just do Join()s with no cleans which is a bit easier to understand but more expensive since we are walking the same path over and over again with an additional segment added.

Patch-set: 1
Reviewer: Gerrit User 5889 <5889@62eb7196-b449-3ce5-99f1-c037f21e1705>