Update patch set 4
Patch Set 4:

Root cause of the check and builder failure is we have more than one VarDef for the same variable in the same block.  Proper scheduling would get this right because the OpAddr is (eventually) an input to one of the store operations, but "StoreOrder" is not proper scheduling.  Though we don't have VarDef/VarKill in this instance (so it is especially questionable) we certainly could, and do for autotmps, so my CL was not good enough anyhow.

I tried just generically scheduling the block instead and for some reason got some other crash.  Happy-happy-joy-joy.  I think this fix is going to end up somewhat more invasive than we're happy with -- what I imagine is a general-purpose scheduler oriented towards clumping stores together.  Any other suggestions would be welcome.

Patch-set: 4