Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> How about as an interim solution we eliminate all but one of the VarDefs?

That would get us past this one, which might be a good idea.  We could try using the first one.

One risk that I see in general is for pointer-typed autotmps -- those seem to get paired with VarKill, if we move an OpAddr from the second (later) def/kill pair into the first def/kill pair, then we have a stack pointer live where it previously was not and where the def/kill information says it isn't supposed to be live.

The other thing I wonder if we should try is to add OpLocalAddr, and make it take a mem, which it gets from the VarDef for the local.  This would require a certain amount of boilerplate work throughout the compiler, also not happy-making (98 hits for "git grep OpAddr", could be worse).  This seems more correct to me.

Patch-set: 4