Update patch set 3
Patch Set 3:

> > What do you think of the discussion at golang.org/cl/111257 (and issue #25201)?
> 
> (Perhaps the argument is that this CL targets only panics that should clearly never happen in the runtime, while leaving things like panicmem alone.)

I couldn't find that CL, so thanks.  Yes, I feel that this is different in that it only targets index and slice expressions.  I can't think of any case in which the runtime package would panic on an index or slice expression and expect user code to catch it.  I also chose to explicitly avoid the confusion about signal handlers and panicmem, panicdivide, etc.

This approach is a little nicer in that it doesn't add extra compiler hacks for the runtime package, a little worse in that it does some extra checks on every index/slice panic.  But I would guess that such panics are rare in production code.

Patch-set: 3
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>