Update patch set 1
Patch Set 1:

(3 comments)

I think this is testing less things now.

Also, it seems a simpler approach might be to prepare a[] as all eights,
and b[] as all nines, not sixes and sevens?, and then have the 'randomish'
data gradually trample them from the start.

The lastLen approach would help cut the work in overwriting the [89],
but the tests for which it has been introduced would seem clearer outside of
this loop, in their own, as it can just be run once at the end when
a[] and b[] are all randomish, with no lastLen housekeeping required.

Patch-set: 1
CC: Gerrit User 10961 <10961@62eb7196-b449-3ce5-99f1-c037f21e1705>