Update patch set 1
Patch Set 1:

Ian, it's been a full week and it would be great to include this fix in the first release candidate. That will give us time to try another approach or revert the original CL if anyone still notices breakage in that release.

Is there anyone we could ping for a +2?

Patch-set: 1
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>