Update patch set 1
Patch Set 1:

To be honest, I'm not sure I agree that calling the Bytes method makes the code more readable.  It means that somebody reading the code has to go look at the Bytes method and see what it does.  The direct references to buf seems more clear to me.

Patch-set: 1
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>