cmd/compile: add mem operand to OpAddr
Lack of a well-defined order between VarDef and related
address operations sometimes causes problems with store order
and write barrier transformations; glitches in the order are
made irreparable (by later optimizations) if the two parts of
the glitch straddle a split in the original block caused by
insertion of a write barrier diamond.

Fix this by adding a memory input to OpAddr to define an order,
and in some cases slightly reorder the VarDef and OpAddr
operations.  The new ordering is only present at the generic
SSA level; it is dropped in translation to the machine-dependent
opcodes.

This removes the opportunity to perform some CSE on OpAddr;
the cost of this is not yet known (benchmarking coming), but
there is also a follow-on CL that fixes this if that is felt
to be necessary.

Fixes #26105.

Change-Id: I9ff526a5f06d4551fe45c63011c59c77f41738a6