Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (1 comment)

I investigated further and updated with a new patch set.

The error code of the close() must be EOF to consider 
that data was valid. This err code is now always returned.

A bool stores in the decoder if this extra byte
is expected. If expected and not found, an error is returned.

The solution is of course a bit more complex than expected but
all 20 images are now accepted.

I added your playground example as test for this case.

Regards,

Patch-set: 2
Reviewer: Gerrit User 26957 <26957@62eb7196-b449-3ce5-99f1-c037f21e1705>