cmd/go: fix stale packages long test
This test was failing on my Linux laptop, in a very similar manner to
how it was failing on the linux-amd64-longtest builder:

	go_test.go:933: ./testgo list claims p1 is NOT stale, incorrectly, after restoring sys.go

This is presumably due to GOCACHE, as setting it to a temporary dir that
changes for each test run makes the issue completely go away on my
machine.

Updates #24436.

Change-Id: I5d4ceb3ec9b42c154a08a95bf0a724f5713c91e9