Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> I think Russ said that the "go" repo is now the canonical repo and "vgo" will be mechanically generated from "go". So I think this is safe to re-submit.
> 
> Did we lose other changes, though?

The previous fix was in fetch_test.go (https://golang.org/cl/120735), whereas this one is in git_test.go.

Presumably the fix was lost because the file was renamed in the vgo repo, at a time when the vgo repository was still the source of truth for the file.


(I diff'd the two repos when reviewing https://golang.org/cl/123576, and I did not see any changes that would affect anything outside of the vgo-related parts of the go command. It is of course possible that I missed something.)

Patch-set: 1
CC: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>