net/http: fix double-close of req.Body after request writing
It is possible for a request body to write successfuly in transfer.go
but fail at flushing in a persistConn's writeLoop.

The successful write closes a request body; the failed flush also closes
a request body.

This borrows on the idea of the transferWriter's bodyReadError field and
adds a postCloseError field. If a transferWriter errors on or after
BodyCloser, postCloseError will be true and we can bubble the fact that
the body is closed up appropriately.

writeLoop now takes into account that a nil error means the body was
closed and that the error being the requestPostBodyClosedError type
also means that the body has been closed. If the writeLoop's flush on
nil error fails, we will avoid double closing.

The test required changing countCloseReader's counter to an atomic.
In the other uses, the request body is definitely read before the
response is received. In the new test, the response can be seen before
the request body is entirely written and closed.

Updates #19186
Updates #26408

Change-Id: I5b7d8741a460d70cb20c577782e3b868b9340331