cmd/compile: keep autos if their address reaches a block control
Autos must be kept and set to zero if their address is used. In most
cases if the address of an auto reaches a control statement it will
also reach another operation that causes it to be kept, such as a phi,
however this is not currently always true.

Fixes #26407.

Change-Id: I02265793f010a9e001c3e1a5397c290c6769d4de