Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> Trybot failures are real.

Oops, I was expecting godoc to warn if that was an issue.  Removed the stray import.

I also added an os.Chdir call so the output is consistent regardless of the particular temp directory that is created (I also didn't realise that the output comment was used to verify the results).

Patch-set: 6
Reviewer: Gerrit User 18980 <18980@62eb7196-b449-3ce5-99f1-c037f21e1705>