cmd/go: fix TestScript on Plan 9
CL 123577 added TestScript. The install_rebuild_gopath
test was failing on Plan 9 because it defines a GOPATH
using the ':' separator, while Plan 9 expects the '\000'
separator in environment variables.

This change fixes the script engine by adding a new "envpath"
command, similar to "env", but replacing the ':' separator
in environment variables by os.PathListSeparator when required.

Fixes #26421.

Change-Id: I58a97f882cdb48cc0836398b0d98a80ea58041ba