Update patch set 2
Patch Set 2:

This text is using terminology that's inconsistent with the actual runtime/trace API. But more generally, do we want to include this much detail in the release notes? At this level, it would seem better just to give a few sentences and point to the package docs, which explains all of this carefully and succinctly.

Patch-set: 2
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>