go/format: add benchmark
Generate a file that resembles the file generated by
github.com/kevinburke/go-bindata, a tool for embedding assets in a Go
binary, that is currently slow to format. Hopefully this can be used
in the future to measure and improve the performance of the Go format
tool.

Results on a Macbook Pro:

    name            time/op
    FormatSource-4     1.28s ± 0%

    name            speed
    FormatSource-4  73.1MB/s ± 0%

    name            alloc/op
    FormatSource-4    1.02GB ± 0%

    name            allocs/op
    FormatSource-4     9.87k ± 0%

Fixes golang/go#26528.

Change-Id: I29fe2bfdfa34ff18ab6b16dea739c6113392dd68