Update patch set 2
Patch Set 2:

> Maybe it's me, but the new version seems significantly less readable
> and harder to understand than the old one (for a speed gain that is
> measurable, but also not huge).

Maybe so. If this was only inlining for the sake of avoiding the call overhead, I'd agree that we should just wait for mid-stack inlining to land.

However, this is about adding a short-circuit to the logic, which I'm not even sure would be within the scope of the compiler to do as an optimization. Even if fieldByIndex and IsValid were inlined, I don't think the compiler could rewrite the program to this form.

Either way, I'm happy to let this sit for a bit and wait to see how mid-stack inlining would affect the performance numbers.

Patch-set: 2
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>