net/http: fix no-close of req.Body after redirect
The fix for #19186 (CL 37298) was incorrect in that, after the first
request, reqBodyClosed was never updated to false on a new request after
a redirect if the request's GetBody returned a new body.

This resets reqBodyClosed after GetBody in a redirect.

Further, this rips the closing-on-error logic from uerr into a defer.
CheckRedirect could return a non-nil error _after_ GetBody was called.
In this case, reqBodyClosed was never checked.

Change-Id: I7043a30882a114a656311f3a9c3c3041e451028d