crypto/x509: allow certificates to be exported from CertPool
Some users may find it helpful to be able to inspect the `Roots`/
`Intermediates` in a `VerifyOptions`.

The main issue with exposing this is that users could mess with the slice,
causing the two indices (`bySubjectKeyId` and `byName`) to be broken. To
avoid this, a `Certificates()` method is added, returning a copy of the
internal slice.

Change-Id: Ib3107ab99d7d031bae451e734923d4c72c44d398
GitHub-Last-Rev: 0846e57b50fdd5a6768c10feea9053a3c7583484
GitHub-Pull-Request: golang/go#26613