Update patch set 1
Patch Set 1: Code-Review-2

Hi! Thanks for contributing.

API changes need to be discussed in a Proposal (it's a lightweight process, it just takes opening an issue with a title starting with "proposal:").

We will have to be careful considering this change because we don't want to necessarily require the availability of full certificates on all OSes to support SystemCertPool, as opposed to just the subjects which might be all the system API returns.

Patch-set: 1
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2